refactored to avoid duplicate code
